enhance: move import file probing into the format packages
Sizing needed the row count a parquet footer and an npy header record, and
that probing was written in datacoord -- which made it the only package
importing npyio and arrow's parquet reader, and restated format knowledge
the reader packages already owned. numpy.SourcePaths was extracted earlier
in this PR for exactly that reason; this applies the same rule to the rest.

RowCountUpperBound now dispatches per format next to NewReader, and each
format package holds its own probe. The footer decode gate stays attached
to the sizing entry point, not to NewReader, so DataNode reads are
unaffected. Error and log text is unchanged.

issue: #51667
