+++ internal/datacoord/import_row_bound.go
@@ -17,490 +17,22 @@
 package datacoord
 
 import (
-	"bytes"
 	"context"
-	"encoding/binary"
-	"fmt"
-	"io"
 	"math"
-	"time"
-
-	"github.com/apache/arrow/go/v17/parquet/file"
-	"github.com/cockroachdb/errors"
-	"github.com/sbinet/npyio"
 
 	"github.com/milvus-io/milvus-proto/go-api/v3/schemapb"
 	"github.com/milvus-io/milvus/internal/storage"
 	"github.com/milvus-io/milvus/internal/util/importutilv2"
-	importutilv2common "github.com/milvus-io/milvus/internal/util/importutilv2/common"
-	"github.com/milvus-io/milvus/internal/util/importutilv2/numpy"
 	"github.com/milvus-io/milvus/pkg/v3/common"
-	"github.com/milvus-io/milvus/pkg/v3/mlog"
 	"github.com/milvus-io/milvus/pkg/v3/proto/internalpb"
 	"github.com/milvus-io/milvus/pkg/v3/util/conc"
 	"github.com/milvus-io/milvus/pkg/v3/util/hardware"
 	"github.com/milvus-io/milvus/pkg/v3/util/merr"
 	"github.com/milvus-io/milvus/pkg/v3/util/paramtable"
-	"github.com/milvus-io/milvus/pkg/v3/util/retry"
-	"github.com/milvus-io/milvus/pkg/v3/util/typeutil"
 )
 
-// nonSourceFieldIDs returns field IDs absent from import source files: the
-// autoID primary key (generated by Milvus) and all function-output fields
-// (generated during import, e.g. a BM25 sparse vector).
-func nonSourceFieldIDs(schema *schemapb.CollectionSchema) typeutil.Set[int64] {
-	ids := typeutil.NewSet[int64]()
-	for _, f := range schema.GetFields() {
-		if f.GetIsPrimaryKey() && f.GetAutoID() {
-			ids.Insert(f.GetFieldID())
-		}
-		// The dynamic field is never required from the source file: the JSON row
-		// parser leaves it out of name2FieldID (json/row_parser.go:80-83), so the
-		// completeness check never asks for it and combineDynamicRow synthesizes
-		// it from the row's spare keys. The nullable/default skip below covers it
-		// only for collections created after $meta gained those attributes; a
-		// schema persisted before that is neither, and nothing migrated it.
-		if f.GetIsDynamic() {
-			ids.Insert(f.GetFieldID())
-		}
-	}
-	for _, fn := range schema.GetFunctions() {
-		ids.Insert(fn.GetOutputFieldIds()...)
-	}
-	return ids
-}
-
-// minRowTextBytes returns a provable lower bound on the number of bytes one row
-// occupies in a text import file of type ft, EXCLUDING fields not present in the
-// source files (autoID primary key and function-output fields).
-//
-// Two parts contribute. The per-field value floor is format-independent: a known
-// dense vector needs at least one character per element (dim, or dim/8 for binary
-// vectors); every other known fixed scalar needs at least 1; VarChar/JSON may be
-// empty and unknown/sparse source fields need none.
-//
-// The structural floor is NOT format-independent, and it is what keeps a large
-// all-VarChar file from being estimated at one row per byte: a JSON row spells out
-// every field name it carries, paying {"name": per field, whereas a CSV row carries
-// only separators because its names live in the header.
-//
-// Row separators are deliberately not counted: a single-row file has none, so
-// counting one would let the floor exceed a real row and under-count the file --
-// the one direction this bound may never take. The result is floored to 1 to avoid
-// divide-by-zero downstream.
-func minRowTextBytes(schema *schemapb.CollectionSchema, ft importutilv2.FileType) int64 {
-	// A JSON row is an object: braces around it, and "name": before each value.
-	jsonShaped := ft == importutilv2.JSON || ft == importutilv2.JSONLines
-	skip := nonSourceFieldIDs(schema)
-	var total, present int64
-	for _, field := range schema.GetFields() {
-		if skip.Contain(field.GetFieldID()) {
-			continue
-		}
-		// A nullable or defaulted field may be omitted entirely from a JSON row
-		// (the reader fills it), contributing 0 bytes. Counting it would overstate
-		// the per-row floor and understate the row count, so skip it to keep this a
-		// true lower bound.
-		if field.GetNullable() || field.GetDefaultValue() != nil {
-			continue
-		}
-		present++
-		if jsonShaped {
-			// Two quotes and a colon around the field name.
-			total += int64(len(field.GetName())) + 3
-		}
-		switch field.GetDataType() {
-		case schemapb.DataType_Bool, schemapb.DataType_Int8, schemapb.DataType_Int16,
-			schemapb.DataType_Int32, schemapb.DataType_Int64,
-			schemapb.DataType_Float, schemapb.DataType_Double:
-			// Known fixed scalar: at least 1 character.
-			total++
-		case schemapb.DataType_FloatVector, schemapb.DataType_Float16Vector,
-			schemapb.DataType_BFloat16Vector, schemapb.DataType_Int8Vector:
-			dim, err := typeutil.GetDim(field)
-			if err != nil {
-				// Unknown dimension: cannot prove a floor, contributes 0.
-				continue
-			}
-			total += dim
-		case schemapb.DataType_BinaryVector:
-			dim, err := typeutil.GetDim(field)
-			if err != nil {
-				continue
-			}
-			total += dim / 8
-		default:
-			// VarChar/JSON may be empty; unknown/sparse source fields contribute 0.
-		}
-	}
-	if present > 1 {
-		// One separator between adjacent fields: ',' in both formats.
-		total += present - 1
-	}
-	if jsonShaped {
-		total += 2 // the row object's braces
-	}
-	if total < 1 {
-		total = 1
-	}
-	return total
-}
-
-// computeFileRowUpperBound returns an upper bound on the row count of one
-// ImportFile, and whether that bound is exact. Parquet and numpy record their row
-// count in the file itself (footer num_rows / header shape), so both are exact.
-// JSON/CSV record nothing, so their bound divides Σ file size by a provable
-// per-row byte floor -- a heavy over-estimate that is still guaranteed not to
-// under-count. The exactness decides how the caller sizes the reservation and,
-// under ID pressure, which reservations it may shrink.
-func computeFileRowUpperBound(ctx context.Context, cm storage.ChunkManager,
-	schema *schemapb.CollectionSchema, file *internalpb.ImportFile,
-) (int64, bool, error) {
-	ft, err := importutilv2.GetFileType(file)
-	if err != nil {
-		return 0, false, err
-	}
-
-	sumSize := func() (int64, error) {
-		var total int64
-		for _, path := range file.GetPaths() {
-			size, err := cm.Size(ctx, path)
-			if err != nil {
-				return 0, err
-			}
-			total += size
-		}
-		return total, nil
-	}
-
-	var bound int64
-	var exact bool
-	switch ft {
-	case importutilv2.Parquet:
-		bound, err = parquetNumRows(ctx, cm, file.GetPaths()[0])
-		if err != nil {
-			return 0, false, err
-		}
-		exact = true
-	case importutilv2.Numpy:
-		bound, err = numpyNumRows(ctx, cm, schema, file.GetPaths())
-		if err != nil {
-			return 0, false, err
-		}
-		exact = true
-	case importutilv2.JSON, importutilv2.CSV, importutilv2.JSONLines:
-		// No row count is recorded in the file, so divide the byte size by a
-		// provable per-row floor. This over-estimates heavily (the floor is 1 byte
-		// for an all-VarChar schema) and must stay an upper bound: under-estimating
-		// would exhaust the range and fail the import at the datanode guard.
-		minRow := minRowTextBytes(schema, ft)
-		total, err := sumSize()
-		if err != nil {
-			return 0, false, err
-		}
-		bound = total/minRow + 1
-	default:
-		return 0, false, merr.WrapErrImportFailed("unknown import file type for PK range sizing")
-	}
-
-	if bound < 0 {
-		bound = 0
-	}
-	return bound, exact, nil
-}
-
-// maxNpyHeaderLen caps the header length a .npy file may declare. numpy pads the
-// header to a 64-byte boundary and it carries a single dtype/shape dict, so real
-// headers are a few hundred bytes; a v1 length is a uint16 and can never exceed
-// this bound at all.
-const maxNpyHeaderLen = 1 << 16
-
-// validateNpyHeader checks the preconditions npyio's readHeader assumes but never
-// verifies (npy/reader.go:97-101): it allocates the declared header length
-// verbatim, then slices on the last '\n' without checking r.err or that a newline
-// exists. So a v2 file declaring a 4 GiB header allocates 4 GiB per sizing
-// goroutine, and a header carrying no newline yields index -1 and panics. Both run
-// on the coordinator here, where a panic is not recoverable by the caller: the
-// sizing pool's worker is a separate goroutine.
-//
-// Only the prefix is read, via ReadAt, so the reader's sequential offset is
-// untouched and npyio still decodes from the start.
-func validateNpyHeader(ra io.ReaderAt, path string) error {
-	var prefix [12]byte
-	n, err := ra.ReadAt(prefix[:], 0)
-	if err != nil && !errors.Is(err, io.EOF) {
-		return err
-	}
-	if n < 10 || [6]byte(prefix[0:6]) != npyio.Magic {
-		return merr.WrapErrImportFailedMsg("not a numpy file, path=%s", path)
-	}
-	var hdrLen, hdrOff int64
-	switch major := prefix[6]; major {
-	case 1:
-		hdrLen, hdrOff = int64(binary.LittleEndian.Uint16(prefix[8:10])), 10
-	case 2:
-		if n < 12 {
-			return merr.WrapErrImportFailedMsg("numpy v2 header truncated, path=%s", path)
-		}
-		hdrLen, hdrOff = int64(binary.LittleEndian.Uint32(prefix[8:12])), 12
-	default:
-		return merr.WrapErrImportFailedMsg("unsupported numpy major version %d, path=%s", major, path)
-	}
-	if hdrLen <= 0 || hdrLen > maxNpyHeaderLen {
-		return merr.WrapErrImportFailedMsg("numpy header length %d out of range (max %d), path=%s",
-			hdrLen, maxNpyHeaderLen, path)
-	}
-	hdr := make([]byte, hdrLen)
-	if _, err := ra.ReadAt(hdr, hdrOff); err != nil && !errors.Is(err, io.EOF) {
-		return err
-	}
-	if bytes.IndexByte(hdr, '\n') < 0 {
-		return merr.WrapErrImportFailedMsg("numpy header is not newline-terminated, path=%s", path)
-	}
-	return nil
-}
-
-// numpyNumRows reads only the .npy headers to get the exact row count of one
-// column-based import file. Every path belongs to the same rows, so their shapes
-// agree on a well-formed input; the maximum is taken so a disagreement can only
-// over-reserve rather than under-reserve (the reader rejects the mismatch later).
-// Paths for omitted nullable/default columns are simply absent and contribute
-// nothing.
-//
-// Only the paths the reader will open are inspected. Sizing runs before the
-// broadcast, so validating a path the reader ignores would reject at submit an
-// input that imports fine -- a redundant .npy naming no schema field is dropped
-// by numpy.CreateReaders, and must be dropped here by the same rule. When that
-// leaves no path at all, this returns 0 rather than an error: the file carries no
-// readable column, and saying so is the reader's job, at the same place it said
-// so before this sizing stage existed.
-func numpyNumRows(ctx context.Context, cm storage.ChunkManager, schema *schemapb.CollectionSchema, paths []string) (int64, error) {
-	var rows int64
-	for _, path := range numpy.SourcePaths(schema, paths) {
-		ra, err := newSizingReaderAt(ctx, cm, path)
-		if err != nil {
-			return 0, err
-		}
-		if err := validateNpyHeader(ra, path); err != nil {
-			return 0, err
-		}
-		r, err := npyio.NewReader(ra)
-		if err != nil {
-			return 0, importutilv2common.WrapDecodeErr(err, fmt.Sprintf("read numpy header failed, path=%s", path))
-		}
-		shape := r.Header.Descr.Shape
-		if len(shape) == 0 {
-			continue
-		}
-		if n := int64(shape[0]); n > rows {
-			rows = n
-		}
-	}
-	return rows, nil
-}
-
-var (
-	parquetMagic          = []byte("PAR1")
-	parquetMagicEncrypted = []byte("PARE")
-)
-
-// parquetFooterMaxSize returns the largest footer metadata length import sizing
-// will read, from dataCoord.import.parquetFooterMaxSize.
-//
-// Arrow reads the declared length from the file's last 8 bytes
-// (parquet/file/file_reader.go:174) and allocates it verbatim (:182, :202), and
-// sizingReaderAt.ReadAt buffers a whole ranged GET before copying, so a parse
-// costs about twice the declared length.
-//
-// This is stricter than the DataNode reader, which bounds the footer only by the
-// file's own size, so a file above the limit is refused at submit although the
-// reader would have read it. Footer size tracks row_groups * columns: 128 MiB row
-// groups under Milvus's default 64-field ceiling land in the low single-digit MiB,
-// but small row groups, wide schemas or untruncated string statistics do not, which
-// is why the limit is configurable rather than fixed.
-func parquetFooterMaxSize() int64 {
-	v := paramtable.Get().DataCoordCfg.ImportParquetFooterMaxSize.GetAsInt64()
-	if v <= 0 {
-		return defaultParquetFooterMaxSize
-	}
-	return v
-}
-
-// defaultParquetFooterMaxSize mirrors the paramtable default; used when the
-// configured value is missing or nonsensical.
-const defaultParquetFooterMaxSize = 64 << 20
-
-// maxConcurrentParquetFooterParses bounds how many parquet footers are parsed at
-// once across the whole process.
-//
-// parquetFooterMaxSize bounds the bytes read, not the memory the decode takes:
-// Arrow's generated reader sizes []*RowGroup from the element count the footer
-// declares (parquet/internal/gen-go/parquet/parquet.go:12252) before reading a
-// single element, while thrift's guard compares that count against a *byte*
-// limit (thrift/configuration.go:305, default 100 MiB). A 6-byte footer may
-// therefore declare 104857600 elements and allocate 800 MiB of pointers. The cap
-// is process-wide, not per-request, because every import request builds its own
-// sizing pool and nothing above them limits how many run at once.
-const maxConcurrentParquetFooterParses = 4
-
-var parquetFooterParseSem = make(chan struct{}, maxConcurrentParquetFooterParses)
-
-// parquetFooterGateSlowWait is how long a sizing pass may wait for the gate
-// before the wait is worth a log line.
-const parquetFooterGateSlowWait = 5 * time.Second
-
-// validateParquetFooter rejects an out-of-range declared footer length before Arrow
-// allocates it. sizingReaderAt.ReadAt buffers a whole ranged GET before copying, so
-// each parse costs about twice the declared length.
-//
-// This bounds the bytes one file may read, which is not the same as the memory its
-// decode takes -- see maxConcurrentParquetFooterParses, which bounds how many parses
-// run at once and is what actually caps the coordinator's exposure. A per-pass byte
-// budget is still absent; the concurrency cap stands in for it.
-func validateParquetFooter(ra io.ReaderAt, size int64, path string) error {
-	// Leading magic (4) + metadata length (4) + trailing magic (4).
-	const minParquetSize = 12
-	if size < minParquetSize {
-		return merr.WrapErrImportFailedMsg("parquet file too small, size=%d, path=%s", size, path)
-	}
-	var tail [8]byte
-	if _, err := ra.ReadAt(tail[:], size-int64(len(tail))); err != nil && !errors.Is(err, io.EOF) {
-		return err
-	}
-	if !bytes.Equal(tail[4:], parquetMagic) && !bytes.Equal(tail[4:], parquetMagicEncrypted) {
-		return merr.WrapErrImportFailedMsg("not a parquet file, path=%s", path)
-	}
-	footerLen := int64(binary.LittleEndian.Uint32(tail[:4]))
-	if maxLen := parquetFooterMaxSize(); footerLen <= 0 || footerLen > maxLen {
-		return merr.WrapErrImportFailedMsg(
-			"parquet footer length %d out of range (max %d, dataCoord.import.parquetFooterMaxSize), path=%s",
-			footerLen, maxLen, path)
-	}
-	return nil
-}
-
-// parquetNumRows reads only the parquet footer to get the exact row count via a
-// sizingReaderAt, which retries the ranged reads the footer relies on. Any error
-// from NewParquetReader is therefore a genuine file-format problem, not a transient
-// fault, and is returned as a non-retryable import error.
-func parquetNumRows(ctx context.Context, cm storage.ChunkManager, path string) (int64, error) {
-	ra, err := newSizingReaderAt(ctx, cm, path)
-	if err != nil {
-		return 0, err
-	}
-	if err := validateParquetFooter(ra, ra.size, path); err != nil {
-		return 0, err
-	}
-
-	// Hold the gate across the decode only. The reads above are ordinary object
-	// storage traffic with their own retry/backoff, and holding a global slot
-	// through them would flatten the caller's pool to this gate's width and let
-	// one slow request stall every other import's sizing pass.
-	//
-	// The gate is process-wide and narrow, so concurrent submissions of
-	// footer-heavy files queue behind each other and lengthen the broadcast RPC.
-	// That is the intended trade against unbounded decode allocation, but it is
-	// invisible from the outside, so a wait worth noticing is logged.
-	gateStart := time.Now()
-	select {
-	case parquetFooterParseSem <- struct{}{}:
-	case <-ctx.Done():
-		return 0, ctx.Err()
-	}
-	if waited := time.Since(gateStart); waited > parquetFooterGateSlowWait {
-		mlog.Warn(ctx, "waited for the parquet footer decode gate",
-			mlog.String("path", path),
-			mlog.Duration("waited", waited),
-			mlog.Int("gateWidth", maxConcurrentParquetFooterParses))
-	}
-	defer func() { <-parquetFooterParseSem }()
-
-	pr, err := file.NewParquetReader(ra)
-	if err != nil {
-		return 0, importutilv2common.WrapDecodeErr(err, fmt.Sprintf("read parquet footer failed, path=%s", path))
-	}
-	defer pr.Close()
-	return pr.NumRows(), nil
-}
-
-// sizingReaderAt is a minimal io.Reader / io.ReaderAt / io.Seeker over a
-// ChunkManager, used only to read the small header/footer a sizing pass needs.
-// Each ReadAt is an independent ranged GET wrapped in retry, so a brief
-// object-store fault is absorbed at the fetch layer. retry.Handle gives up after a
-// bounded number of attempts though, so a persistent outage does surface to the
-// decoder as a typed IO error — see wrapSizingErr, which is why a decoder failure
-// may not be assumed to be a file-format problem. Unlike RetryableReader it retries the
-// ReadAt/Seek path the parquet footer relies on, and it holds no open handle
-// (nothing to close, no reopen-at-offset logic).
-type sizingReaderAt struct {
-	ctx  context.Context
-	cm   storage.ChunkManager
-	path string
-	size int64
-	off  int64
-}
-
-func newSizingReaderAt(ctx context.Context, cm storage.ChunkManager, path string) (*sizingReaderAt, error) {
-	size, err := cm.Size(ctx, path) // Size retries internally.
-	if err != nil {
-		return nil, storage.ToMilvusIoError(path, err)
-	}
-	return &sizingReaderAt{ctx: ctx, cm: cm, path: path, size: size}, nil
-}
-
-func (r *sizingReaderAt) ReadAt(p []byte, off int64) (int, error) {
-	if off < 0 {
-		return 0, merr.WrapErrParameterInvalidMsg("negative read offset %d", off)
-	}
-	if len(p) == 0 || off >= r.size {
-		if off >= r.size {
-			return 0, io.EOF
-		}
-		return 0, nil
-	}
-	var data []byte
-	if err := retry.Handle(r.ctx, func() (bool, error) {
-		b, e := r.cm.ReadAt(r.ctx, r.path, off, int64(len(p)))
-		if e != nil {
-			ioErr := storage.ToMilvusIoError(r.path, e)
-			return !merr.IsNonRetryableErr(ioErr), ioErr
-		}
-		data = b
-		return false, nil
-	}); err != nil {
-		return 0, err
-	}
-	n := copy(p, data)
-	if n < len(p) {
-		return n, io.EOF
-	}
-	return n, nil
-}
-
-func (r *sizingReaderAt) Read(p []byte) (int, error) {
-	n, err := r.ReadAt(p, r.off)
-	r.off += int64(n)
-	return n, err
-}
-
-func (r *sizingReaderAt) Seek(offset int64, whence int) (int64, error) {
-	var abs int64
-	switch whence {
-	case io.SeekStart:
-		abs = offset
-	case io.SeekCurrent:
-		abs = r.off + offset
-	case io.SeekEnd:
-		abs = r.size + offset
-	default:
-		return 0, merr.WrapErrParameterInvalidMsg("invalid seek whence %d", whence)
-	}
-	if abs < 0 {
-		return 0, merr.WrapErrParameterInvalidMsg("negative seek position %d", abs)
-	}
-	r.off = abs
-	return abs, nil
-}
+// maxIDsPerAllocBatch mirrors the per-call ceiling of rootCoordAllocator.AllocN.
+const maxIDsPerAllocBatch = int64(math.MaxUint32)
 
 // assignPKRangesToFiles computes a per-file row upper bound, allocates one
 // contiguous primary-namespace id block sized Σbound via allocN, then writes each
@@ -523,58 +55,6 @@ func assignPKRangesToFiles(ctx context.Context, cm storage.ChunkManager,
 	return reserveRanges(bounds, files, allocN, clusterID)
 }
 
-// maxIDsPerAllocBatch mirrors the per-call ceiling of rootCoordAllocator.AllocN.
-const maxIDsPerAllocBatch = int64(math.MaxUint32)
-
-// reserveRanges writes each file its own contiguous [PkIdBegin, PkIdEnd) slice,
-// chunking the allocation so that no single allocN call exceeds the allocator's
-// per-batch ceiling. Files are packed greedily and a file's range never straddles
-// two batches, so every range stays contiguous while the import as a whole is not
-// capped at one batch: the reservation total may exceed the ceiling, only a single
-// file may not.
-func reserveRanges(bounds []int64, files []*internalpb.ImportFile,
-	allocN func(int64) (int64, int64, error), clusterID uint64,
-) error {
-	for i := 0; i < len(bounds); {
-		var batch int64
-		j := i
-		for ; j < len(bounds); j++ {
-			if bounds[j] > maxIDsPerAllocBatch {
-				// For a format that records no row count the number below is an upper
-				// bound derived from the file size, so it can exceed the ceiling on a
-				// file holding far fewer rows. Say so, or the operator reads it as a
-				// real row count and has nowhere to go.
-				return merr.WrapErrParameterInvalidMsg(
-					"import file %d needs up to %d primary keys, more than one allocation batch holds (max %d); "+
-						"for json/csv this is an upper bound derived from the file size -- split the file, "+
-						"or use parquet/numpy, which record their row count exactly",
-					j, bounds[j], maxIDsPerAllocBatch)
-			}
-			if batch+bounds[j] > maxIDsPerAllocBatch {
-				break
-			}
-			batch += bounds[j]
-		}
-		if batch == 0 {
-			// Only reachable once every remaining bound is zero, which
-			// sizeReservations rules out; guards against a non-advancing loop.
-			return nil
-		}
-		begin, _, err := common.AllocAutoIDN(allocN, batch, clusterID)
-		if err != nil {
-			return err
-		}
-		cur := begin
-		for k := i; k < j; k++ {
-			files[k].PkIdBegin = cur
-			files[k].PkIdEnd = cur + bounds[k]
-			cur = files[k].PkIdEnd
-		}
-		i = j
-	}
-	return nil
-}
-
 // computeFileRowUpperBounds sizes every file concurrently. Sizing is object-store
 // IO (a HEAD per path, or a footer/header read) on the import broadcast path, and
 // the request may carry up to dataCoord.maxFilesPerImportReq files, so doing it
@@ -587,7 +67,7 @@ func computeFileRowUpperBounds(ctx context.Context, cm storage.ChunkManager,
 	// Conceal panics so a decoder crash fails this import instead of the process.
 	// conc.Submit's recover already stores the panic in the future before
 	// re-throwing; concealing stops ants from re-panicking on a worker goroutine,
-	// which no caller here could recover. validateNpyHeader covers the one decoder
+	// which no caller here could recover. The format packages guard the decoder
 	// panic known today -- this covers the ones parquet or a future npyio has left.
 	pool := conc.NewPool[struct{}](hardware.GetCPUNum()*2, conc.WithConcealPanic(true))
 	defer pool.Release()
@@ -596,7 +76,7 @@ func computeFileRowUpperBounds(ctx context.Context, cm storage.ChunkManager,
 	for i, f := range files {
 		i, f := i, f
 		futures = append(futures, pool.Submit(func() (struct{}, error) {
-			b, exact, err := computeFileRowUpperBound(ctx, cm, schema, f)
+			b, exact, err := importutilv2.RowCountUpperBound(ctx, cm, schema, f)
 			if err != nil {
 				return struct{}{}, err
 			}
@@ -658,6 +138,55 @@ func sizeReservations(bounds []int64, exacts []bool) error {
 			b = 1
 		}
 		bounds[i] = b
+	}
+	return nil
+}
+
+// reserveRanges writes each file its own contiguous [PkIdBegin, PkIdEnd) slice,
+// chunking the allocation so that no single allocN call exceeds the allocator's
+// per-batch ceiling. Files are packed greedily and a file's range never straddles
+// two batches, so every range stays contiguous while the import as a whole is not
+// capped at one batch: the reservation total may exceed the ceiling, only a single
+// file may not.
+func reserveRanges(bounds []int64, files []*internalpb.ImportFile,
+	allocN func(int64) (int64, int64, error), clusterID uint64,
+) error {
+	for i := 0; i < len(bounds); {
+		var batch int64
+		j := i
+		for ; j < len(bounds); j++ {
+			if bounds[j] > maxIDsPerAllocBatch {
+				// For a format that records no row count the number below is an upper
+				// bound derived from the file size, so it can exceed the ceiling on a
+				// file holding far fewer rows. Say so, or the operator reads it as a
+				// real row count and has nowhere to go.
+				return merr.WrapErrParameterInvalidMsg(
+					"import file %d needs up to %d primary keys, more than one allocation batch holds (max %d); "+
+						"for json/csv this is an upper bound derived from the file size -- split the file, "+
+						"or use parquet/numpy, which record their row count exactly",
+					j, bounds[j], maxIDsPerAllocBatch)
+			}
+			if batch+bounds[j] > maxIDsPerAllocBatch {
+				break
+			}
+			batch += bounds[j]
+		}
+		if batch == 0 {
+			// Only reachable once every remaining bound is zero, which
+			// sizeReservations rules out; guards against a non-advancing loop.
+			return nil
+		}
+		begin, _, err := common.AllocAutoIDN(allocN, batch, clusterID)
+		if err != nil {
+			return err
+		}
+		cur := begin
+		for k := i; k < j; k++ {
+			files[k].PkIdBegin = cur
+			files[k].PkIdEnd = cur + bounds[k]
+			cur = files[k].PkIdEnd
+		}
+		i = j
 	}
 	return nil
 }
