+++ internal/util/importutilv2/row_count.go
@@ -1,0 +1,197 @@
+// Licensed to the LF AI & Data foundation under one
+// or more contributor license agreements. See the NOTICE file
+// distributed with this work for additional information
+// regarding copyright ownership. The ASF licenses this file
+// to you under the Apache License, Version 2.0 (the
+// "License"); you may not use this file except in compliance
+// with the License. You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+package importutilv2
+
+import (
+	"context"
+
+	"github.com/milvus-io/milvus-proto/go-api/v3/schemapb"
+	"github.com/milvus-io/milvus/internal/storage"
+	"github.com/milvus-io/milvus/internal/util/importutilv2/numpy"
+	"github.com/milvus-io/milvus/internal/util/importutilv2/parquet"
+	"github.com/milvus-io/milvus/pkg/v3/proto/internalpb"
+	"github.com/milvus-io/milvus/pkg/v3/util/merr"
+	"github.com/milvus-io/milvus/pkg/v3/util/typeutil"
+)
+
+// nonSourceFieldIDs returns field IDs absent from import source files: the
+// autoID primary key (generated by Milvus) and all function-output fields
+// (generated during import, e.g. a BM25 sparse vector).
+func nonSourceFieldIDs(schema *schemapb.CollectionSchema) typeutil.Set[int64] {
+	ids := typeutil.NewSet[int64]()
+	for _, f := range schema.GetFields() {
+		if f.GetIsPrimaryKey() && f.GetAutoID() {
+			ids.Insert(f.GetFieldID())
+		}
+		// The dynamic field is never required from the source file: the JSON row
+		// parser leaves it out of name2FieldID (json/row_parser.go:80-83), so the
+		// completeness check never asks for it and combineDynamicRow synthesizes
+		// it from the row's spare keys. The nullable/default skip below covers it
+		// only for collections created after $meta gained those attributes; a
+		// schema persisted before that is neither, and nothing migrated it.
+		if f.GetIsDynamic() {
+			ids.Insert(f.GetFieldID())
+		}
+	}
+	for _, fn := range schema.GetFunctions() {
+		ids.Insert(fn.GetOutputFieldIds()...)
+	}
+	return ids
+}
+
+// minRowTextBytes returns a provable lower bound on the number of bytes one row
+// occupies in a text import file of type ft, EXCLUDING fields not present in the
+// source files (autoID primary key and function-output fields).
+//
+// Two parts contribute. The per-field value floor is format-independent: a known
+// dense vector needs at least one character per element (dim, or dim/8 for binary
+// vectors); every other known fixed scalar needs at least 1; VarChar/JSON may be
+// empty and unknown/sparse source fields need none.
+//
+// The structural floor is NOT format-independent, and it is what keeps a large
+// all-VarChar file from being estimated at one row per byte: a JSON row spells out
+// every field name it carries, paying {"name": per field, whereas a CSV row carries
+// only separators because its names live in the header.
+//
+// Row separators are deliberately not counted: a single-row file has none, so
+// counting one would let the floor exceed a real row and under-count the file --
+// the one direction this bound may never take. The result is floored to 1 to avoid
+// divide-by-zero downstream.
+func minRowTextBytes(schema *schemapb.CollectionSchema, ft FileType) int64 {
+	// A JSON row is an object: braces around it, and "name": before each value.
+	jsonShaped := ft == JSON || ft == JSONLines
+	skip := nonSourceFieldIDs(schema)
+	var total, present int64
+	for _, field := range schema.GetFields() {
+		if skip.Contain(field.GetFieldID()) {
+			continue
+		}
+		// A nullable or defaulted field may be omitted entirely from a JSON row
+		// (the reader fills it), contributing 0 bytes. Counting it would overstate
+		// the per-row floor and understate the row count, so skip it to keep this a
+		// true lower bound.
+		if field.GetNullable() || field.GetDefaultValue() != nil {
+			continue
+		}
+		present++
+		if jsonShaped {
+			// Two quotes and a colon around the field name.
+			total += int64(len(field.GetName())) + 3
+		}
+		switch field.GetDataType() {
+		case schemapb.DataType_Bool, schemapb.DataType_Int8, schemapb.DataType_Int16,
+			schemapb.DataType_Int32, schemapb.DataType_Int64,
+			schemapb.DataType_Float, schemapb.DataType_Double:
+			// Known fixed scalar: at least 1 character.
+			total++
+		case schemapb.DataType_FloatVector, schemapb.DataType_Float16Vector,
+			schemapb.DataType_BFloat16Vector, schemapb.DataType_Int8Vector:
+			dim, err := typeutil.GetDim(field)
+			if err != nil {
+				// Unknown dimension: cannot prove a floor, contributes 0.
+				continue
+			}
+			total += dim
+		case schemapb.DataType_BinaryVector:
+			dim, err := typeutil.GetDim(field)
+			if err != nil {
+				continue
+			}
+			total += dim / 8
+		default:
+			// VarChar/JSON may be empty; unknown/sparse source fields contribute 0.
+		}
+	}
+	if present > 1 {
+		// One separator between adjacent fields: ',' in both formats.
+		total += present - 1
+	}
+	if jsonShaped {
+		total += 2 // the row object's braces
+	}
+	if total < 1 {
+		total = 1
+	}
+	return total
+}
+
+// RowCountUpperBound returns an upper bound on the row count of one
+// ImportFile, and whether that bound is exact. Parquet and numpy record their row
+// count in the file itself (footer num_rows / header shape), so both are exact.
+// JSON/CSV record nothing, so their bound divides Σ file size by a provable
+// per-row byte floor -- a heavy over-estimate that is still guaranteed not to
+// under-count. The exactness decides how the caller sizes the reservation and,
+// under ID pressure, which reservations it may shrink.
+//
+// The per-format dispatch mirrors NewReader's: whatever knowledge of a format's
+// layout this needs belongs in that format's package, next to the reader whose
+// rules it must agree with.
+func RowCountUpperBound(ctx context.Context, cm storage.ChunkManager,
+	schema *schemapb.CollectionSchema, file *internalpb.ImportFile,
+) (int64, bool, error) {
+	ft, err := GetFileType(file)
+	if err != nil {
+		return 0, false, err
+	}
+
+	sumSize := func() (int64, error) {
+		var total int64
+		for _, path := range file.GetPaths() {
+			size, err := cm.Size(ctx, path)
+			if err != nil {
+				return 0, err
+			}
+			total += size
+		}
+		return total, nil
+	}
+
+	var bound int64
+	var exact bool
+	switch ft {
+	case Parquet:
+		bound, err = parquet.NumRows(ctx, cm, file.GetPaths()[0])
+		if err != nil {
+			return 0, false, err
+		}
+		exact = true
+	case Numpy:
+		bound, err = numpy.NumRows(ctx, cm, schema, file.GetPaths())
+		if err != nil {
+			return 0, false, err
+		}
+		exact = true
+	case JSON, CSV, JSONLines:
+		// No row count is recorded in the file, so divide the byte size by a
+		// provable per-row floor. This over-estimates heavily (the floor is 1 byte
+		// for an all-VarChar schema) and must stay an upper bound: under-estimating
+		// would exhaust the range and fail the import at the datanode guard.
+		minRow := minRowTextBytes(schema, ft)
+		total, err := sumSize()
+		if err != nil {
+			return 0, false, err
+		}
+		bound = total/minRow + 1
+	default:
+		return 0, false, merr.WrapErrImportFailed("unknown import file type for PK range sizing")
+	}
+
+	if bound < 0 {
+		bound = 0
+	}
+	return bound, exact, nil
+}
